+++ app/players/controllers/get_player_career_stats_controller.py
@@ -5,7 +5,7 @@ from typing import ClassVar
 from app.config import settings
 from app.controllers import AbstractController
 
-from ..parsers.player_career_parser import PlayerCareerParser
+from ..parsers.player_career_stats_parser import PlayerCareerStatsParser
 
 
 class GetPlayerCareerStatsController(AbstractController):
@@ -13,5 +13,5 @@ class GetPlayerCareerStatsController(AbstractController):
     statistics of a player without labels, easily explorable
     """
 
-    parser_classes: ClassVar[list] = [PlayerCareerParser]
+    parser_classes: ClassVar[list] = [PlayerCareerStatsParser]
     timeout = settings.career_path_cache_timeout
