+++ tests/players/parsers/conftest.py
@@ -2,7 +2,7 @@ import pytest
 from _pytest.fixtures import SubRequest
 
 from app.players.parsers.player_career_parser import PlayerCareerParser
-from app.players.parsers.player_parser import PlayerParser
+from app.players.parsers.player_career_stats_parser import PlayerCareerStatsParser
 from app.players.parsers.player_stats_summary_parser import PlayerStatsSummaryParser
 from app.players.parsers.search_data_parser import NamecardParser
 
@@ -13,8 +13,8 @@ def namecard_parser(request: SubRequest) -> NamecardParser:
 
 
 @pytest.fixture(scope="package")
-def player_parser(request: SubRequest) -> PlayerParser:
-    return PlayerParser(player_id=request.param)
+def player_career_parser(request: SubRequest) -> PlayerCareerParser:
+    return PlayerCareerParser(player_id=request.param)
 
 
 @pytest.fixture(scope="package")
@@ -23,5 +23,5 @@ def player_stats_summary_parser(request: SubRequest) -> PlayerStatsSummaryParser
 
 
 @pytest.fixture(scope="package")
-def player_career_parser(request: SubRequest) -> PlayerCareerParser:
-    return PlayerCareerParser(player_id=request.param)
+def player_career_stats_parser(request: SubRequest) -> PlayerCareerStatsParser:
+    return PlayerCareerStatsParser(player_id=request.param)
